fix: Check name, error, resource, and service names
