+++ tests/snapshots/tests.integrations.laravel.v5_8.common_scenarios_test.test_scenario_get_with_view.json
@@ -1,0 +1,269 @@
+[[
+  {
+    "name": "laravel.request",
+    "service": "laravel_test_app",
+    "resource": "App\\Http\\Controllers\\CommonSpecsController@simple_view unnamed_route",
+    "trace_id": 0,
+    "span_id": 1,
+    "parent_id": 12604606052329515364,
+    "type": "web",
+    "meta": {
+      "_dd.p.dm": "-0",
+      "_dd.p.tid": "664c88c900000000",
+      "component": "laravel",
+      "http.method": "GET",
+      "http.route": "simple_view",
+      "http.status_code": "200",
+      "http.url": "http://localhost:9999/simple_view?key=value&<redacted>",
+      "laravel.route.action": "App\\Http\\Controllers\\CommonSpecsController@simple_view",
+      "laravel.route.name": "unnamed_route",
+      "runtime-id": "5db5654d-5a9e-4cf0-908b-a51f7ced2ff2",
+      "span.kind": "server"
+    },
+    "metrics": {
+      "_sampling_priority_v1": 1.0
+    }
+  },
+     {
+       "name": "laravel.event.handle",
+       "service": "laravel",
+       "resource": "bootstrapping: Illuminate\\Foundation\\Bootstrap\\LoadEnvironmentVariables",
+       "trace_id": 0,
+       "span_id": 2,
+       "parent_id": 1,
+       "type": "web",
+       "meta": {
+         "_dd.base_service": "laravel_test_app",
+         "component": "laravel"
+       }
+     },
+     {
+       "name": "laravel.event.handle",
+       "service": "laravel",
+       "resource": "bootstrapped: Illuminate\\Foundation\\Bootstrap\\LoadEnvironmentVariables",
+       "trace_id": 0,
+       "span_id": 3,
+       "parent_id": 1,
+       "type": "web",
+       "meta": {
+         "_dd.base_service": "laravel_test_app",
+         "component": "laravel"
+       }
+     },
+     {
+       "name": "laravel.event.handle",
+       "service": "laravel",
+       "resource": "bootstrapping: Illuminate\\Foundation\\Bootstrap\\LoadConfiguration",
+       "trace_id": 0,
+       "span_id": 4,
+       "parent_id": 1,
+       "type": "web",
+       "meta": {
+         "_dd.base_service": "laravel_test_app",
+         "component": "laravel"
+       }
+     },
+     {
+       "name": "laravel.event.handle",
+       "service": "laravel_test_app",
+       "resource": "bootstrapped: Illuminate\\Foundation\\Bootstrap\\LoadConfiguration",
+       "trace_id": 0,
+       "span_id": 5,
+       "parent_id": 1,
+       "type": "web",
+       "meta": {
+         "component": "laravel"
+       }
+     },
+     {
+       "name": "laravel.event.handle",
+       "service": "laravel_test_app",
+       "resource": "bootstrapping: Illuminate\\Foundation\\Bootstrap\\HandleExceptions",
+       "trace_id": 0,
+       "span_id": 6,
+       "parent_id": 1,
+       "type": "web",
+       "meta": {
+         "component": "laravel"
+       }
+     },
+     {
+       "name": "laravel.event.handle",
+       "service": "laravel_test_app",
+       "resource": "bootstrapped: Illuminate\\Foundation\\Bootstrap\\HandleExceptions",
+       "trace_id": 0,
+       "span_id": 7,
+       "parent_id": 1,
+       "type": "web",
+       "meta": {
+         "component": "laravel"
+       }
+     },
+     {
+       "name": "laravel.event.handle",
+       "service": "laravel_test_app",
+       "resource": "bootstrapping: Illuminate\\Foundation\\Bootstrap\\RegisterFacades",
+       "trace_id": 0,
+       "span_id": 8,
+       "parent_id": 1,
+       "type": "web",
+       "meta": {
+         "component": "laravel"
+       }
+     },
+     {
+       "name": "laravel.event.handle",
+       "service": "laravel_test_app",
+       "resource": "bootstrapped: Illuminate\\Foundation\\Bootstrap\\RegisterFacades",
+       "trace_id": 0,
+       "span_id": 9,
+       "parent_id": 1,
+       "type": "web",
+       "meta": {
+         "component": "laravel"
+       }
+     },
+     {
+       "name": "laravel.event.handle",
+       "service": "laravel_test_app",
+       "resource": "bootstrapping: Illuminate\\Foundation\\Bootstrap\\RegisterProviders",
+       "trace_id": 0,
+       "span_id": 10,
+       "parent_id": 1,
+       "type": "web",
+       "meta": {
+         "component": "laravel"
+       }
+     },
+     {
+       "name": "laravel.provider.load",
+       "service": "laravel_test_app",
+       "resource": "Illuminate\\Foundation\\ProviderRepository::load",
+       "trace_id": 0,
+       "span_id": 11,
+       "parent_id": 1,
+       "type": "web",
+       "meta": {
+         "component": "laravel"
+       }
+     },
+     {
+       "name": "laravel.event.handle",
+       "service": "laravel_test_app",
+       "resource": "bootstrapped: Illuminate\\Foundation\\Bootstrap\\RegisterProviders",
+       "trace_id": 0,
+       "span_id": 12,
+       "parent_id": 1,
+       "type": "web",
+       "meta": {
+         "component": "laravel"
+       }
+     },
+     {
+       "name": "laravel.event.handle",
+       "service": "laravel_test_app",
+       "resource": "bootstrapping: Illuminate\\Foundation\\Bootstrap\\BootProviders",
+       "trace_id": 0,
+       "span_id": 13,
+       "parent_id": 1,
+       "type": "web",
+       "meta": {
+         "component": "laravel"
+       }
+     },
+     {
+       "name": "laravel.event.handle",
+       "service": "laravel_test_app",
+       "resource": "bootstrapped: Illuminate\\Foundation\\Bootstrap\\BootProviders",
+       "trace_id": 0,
+       "span_id": 14,
+       "parent_id": 1,
+       "type": "web",
+       "meta": {
+         "component": "laravel"
+       }
+     },
+     {
+       "name": "laravel.event.handle",
+       "service": "laravel_test_app",
+       "resource": "Illuminate\\Routing\\Events\\RouteMatched",
+       "trace_id": 0,
+       "span_id": 15,
+       "parent_id": 1,
+       "type": "web",
+       "meta": {
+         "component": "laravel"
+       }
+     },
+     {
+       "name": "laravel.action",
+       "service": "laravel_test_app",
+       "resource": "simple_view",
+       "trace_id": 0,
+       "span_id": 16,
+       "parent_id": 1,
+       "type": "web",
+       "meta": {
+         "component": "laravel"
+       }
+     },
+        {
+          "name": "laravel.event.handle",
+          "service": "laravel_test_app",
+          "resource": "creating: simple_view",
+          "trace_id": 0,
+          "span_id": 19,
+          "parent_id": 16,
+          "type": "web",
+          "meta": {
+            "component": "laravel"
+          }
+        },
+     {
+       "name": "laravel.view.render",
+       "service": "laravel_test_app",
+       "resource": "simple_view",
+       "trace_id": 0,
+       "span_id": 17,
+       "parent_id": 1,
+       "type": "web",
+       "meta": {
+         "component": "laravel"
+       }
+     },
+        {
+          "name": "laravel.event.handle",
+          "service": "laravel_test_app",
+          "resource": "composing: simple_view",
+          "trace_id": 0,
+          "span_id": 20,
+          "parent_id": 17,
+          "type": "web",
+          "meta": {
+            "component": "laravel"
+          }
+        },
+        {
+          "name": "laravel.view",
+          "service": "laravel_test_app",
+          "resource": "/home/circleci/app/tests/Frameworks/Laravel/Version_5_8/resources/views/simple_view.blade.php",
+          "trace_id": 0,
+          "span_id": 21,
+          "parent_id": 17,
+          "type": "web",
+          "meta": {
+            "component": "laravel"
+          }
+        },
+     {
+       "name": "laravel.event.handle",
+       "service": "laravel_test_app",
+       "resource": "Illuminate\\Foundation\\Http\\Events\\RequestHandled",
+       "trace_id": 0,
+       "span_id": 18,
+       "parent_id": 1,
+       "type": "web",
+       "meta": {
+         "component": "laravel"
+       }
+     }]]
